feat: implement LottoController and separate main logic
